Extracted GamezClient to separate file from base store
